resolve differences between old https://github.com/63n/bino.git
Had an old fix where the initialisation of the Qt window
depends on whether bino is running with equalizer or not,
but we don't know this until we handle the command line args,
so the Qt init had to be moved. Couldn't check in the fix
while the old hack was still out on github. just juggling..